SectionContent enum abstraction compatible with cranelift-wasm
In my haste, the first version of SectionContent I wrote had some
problems:
* an unknown custom section's name was dropped
* a binary reader was not available for any of the recognized custom
  sections

cranelift-wasm defers decoding custom sections, except for the name
section, to the `environ` by passing the raw bytes.

This patch changes all custom sections to fall under a single variant
of SectionContent. It contains 1. the name, 2. a binary reader of the
content, and 3. an Option<CustomSectionContent<'a>>. The
CustomSectionContent enum has the specialized reader for each recognized
custom section.