Update parsing of passive elements/simd memargs
It looks like most of the SIMD instructions which take a `MemArg` don't
need to assert that they're a specific alignment, instead they allow
many alignments (exercised in spec tests).

Additionally the encoding of passive element sections has pretty
radically changed from the original implementation. This commit brings
`wasmparser` up to date with the current version of the spec. The
notable changes here are that the flags field is a bitflags field which
has a bunch of bits indicating various types of payloads that may or may
not be present.